apps/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.4.0"
backoff = "2.2.1"
certifi = "2024.6.2"
exceptiongroup = "1.2.1"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
psycopg2 = "2.9.9"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
pydantic-core = "2.18.4"
python-dotenv = "1.0.1"
sniffio = "1.3.1"
typing-extensions = "4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
